refactor network graph classes and update plotting logic
